+++ lib/shared/src/index.ts
@@ -244,4 +244,13 @@ export * from './token/constants'
 export * from './configuration'
 export * from './mentions/providers/packageMentions'
 export * from './mentions/providers/sourcegraphSearch'
+<<<<<<< HEAD
 export * from './cody-ignore/notification'
+||||||| 75accfa42
+=======
+export {
+    setOpenCtxExtensionAPI,
+    getOpenCtxExtensionAPI,
+    type OpenCtxExtensionAPI,
+} from './context/openctx/api'
+>>>>>>> origin/main