+++ vscode/test/e2e/command-custom.test.ts
@@ -63,7 +63,6 @@ test.extend<ExpectedEvents>({
     // Minimize other sidebar items to make room for the command view,
     // else the test will fail because the Custom Command button is not visible
     await expect(page.getByText('Chat alongside your code, attach files,')).toBeVisible()
-    await page.getByLabel('Natural Language Search (Beta) Section').click()
     await page.getByLabel('Settings & Support Section').click()
     await page.getByLabel('Chats Section').click()
 
@@ -75,7 +74,7 @@ test.extend<ExpectedEvents>({
     await page.getByRole('tab', { name: 'index.html' }).hover()
 
     // Bring the cody sidebar to the foreground
-    await page.click('.badge[aria-label="Cody"]')
+    await page.getByRole('tab', { name: 'Cody', exact: true }).locator('a').click()
     // Click the Custom Commands button in the Sidebar to open the Custom Commands menu
     await page.getByText('Custom Commands', { exact: true }).click()
 
@@ -148,7 +147,7 @@ test.extend<ExpectedEvents>({
     await expect(page.getByText(commandName)).toBeVisible()
 
     // Show the new command in the menu and execute it
-    await page.click('.badge[aria-label="Cody"]')
+    await page.getByRole('tab', { name: 'Cody', exact: true }).locator('a').click()
     await openCustomCommandMenu(page)
     await page.getByText('Cody: Custom Commands (Beta)').hover()
     await expect(page.getByText('Cody: Custom Commands (Beta)')).toBeVisible()
@@ -196,7 +195,6 @@ test.extend<ExpectedEvents>({
     // Minimize other sidebar items to make room for the command view,
     // else the test will fail because the Custom Command button is not visible
     await expect(page.getByText('Chat alongside your code, attach files,')).toBeVisible()
-    await page.getByLabel('Natural Language Search (Beta) Section').click()
     await page.getByLabel('Settings & Support Section').click()
     await page.getByLabel('Chats Section').click()
 
@@ -208,7 +206,7 @@ test.extend<ExpectedEvents>({
 
     // Open the chat sidebar to click on the Custom Command option
     // Search for the command defined in cody.json and execute it
-    await page.click('.badge[aria-label="Cody"]')
+    await page.getByRole('tab', { name: 'Cody', exact: true }).locator('a').click()
     await openCustomCommandMenu(page)
 
     /* Test: context.currentDir with currentDir command */
@@ -307,7 +305,6 @@ test.extend<ExpectedEvents>({
     // Minimize other sidebar items to make room for the command view,
     // else the test will fail because the Custom Command button is not visible
     await expect(page.getByText('Chat alongside your code, attach files,')).toBeVisible()
-    await page.getByLabel('Natural Language Search (Beta) Section').click()
     await page.getByLabel('Settings & Support Section').click()
     await page.getByLabel('Chats Section').click()
 
@@ -318,7 +315,7 @@ test.extend<ExpectedEvents>({
     await page.getByRole('treeitem', { name: 'cody.json' }).locator('a').dblclick()
     await page.getByRole('tab', { name: 'cody.json' }).hover()
 
-    await page.click('.badge[aria-label="Cody"]')
+    await page.getByRole('tab', { name: 'Cody', exact: true }).locator('a').click()
     await openCustomCommandMenu(page)
 
     // Able to open the cody.json file in the editor from the command menu
@@ -354,7 +351,7 @@ test.extend<ExpectedEvents>({
 
     // NOTE: This is expected to fail locally if you currently have User commands configured
     await page.waitForTimeout(100)
-    await page.click('.badge[aria-label="Cody"]')
+    await page.getByRole('tab', { name: 'Cody', exact: true }).locator('a').click()
     await openCustomCommandMenu(page)
     await page.locator('a').filter({ hasText: 'Open User Settings (JSON)' }).hover()
     await page.getByRole('button', { name: 'Open or Create Settings File' }).hover()
@@ -388,7 +385,5 @@ testGitWorkspace('use terminal output as context', async ({ page, sidebar }) => 
     await expectContextCellCounts(contextCell, { files: 2 })
     await contextCell.click()
     const chatContext = panel.locator('details').last()
-    await expect(
-        chatContext.getByRole('link', { name: withPlatformSlashes('/terminal-output') })
-    ).toBeVisible()
+    await expect(chatContext.getByRole('link', { name: withPlatformSlashes('/git diff') })).toBeVisible()
 })